Fix SpecimenBatch registration. Move overview and download actions to specimen module.
