+++ specimen/src/org/labkey/specimen/SpecimenModule.java
@@ -26,9 +26,11 @@ import org.labkey.api.module.ModuleContext;
 import org.labkey.api.module.SpringModule;
 import org.labkey.api.pipeline.PipelineService;
 import org.labkey.api.security.roles.RoleManager;
+import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimensPage;
 import org.labkey.api.specimen.importer.SpecimenImporter;
+import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.specimen.model.SpecimenRequestEventType;
 import org.labkey.api.specimen.view.SpecimenRequestNotificationEmailTemplate;
 import org.labkey.api.study.SpecimenService;
@@ -36,10 +38,12 @@ import org.labkey.api.study.StudyService;
 import org.labkey.api.study.importer.SimpleStudyImporterRegistry;
 import org.labkey.api.study.writer.SimpleStudyWriterRegistry;
 import org.labkey.api.util.emailTemplate.EmailTemplateService;
+import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.WebPartFactory;
 import org.labkey.specimen.action.SpecimenApiController;
 import org.labkey.specimen.action.SpecimenController2;
+import org.labkey.specimen.action.SpecimenController2.OverviewAction;
 import org.labkey.specimen.importer.DefaultSpecimenImportStrategyFactory;
 import org.labkey.specimen.importer.SpecimenSchemaImporter;
 import org.labkey.specimen.importer.SpecimenSettingsImporter;
@@ -111,7 +115,22 @@ public class SpecimenModule extends SpringModule
         // service if it's available
         SpecimenService.setInstance(new SpecimenServiceImpl());
         EmailTemplateService.get().registerTemplate(SpecimenRequestNotificationEmailTemplate.class);
-    }
+
+        SpecimenMigrationService.setInstance(new SpecimenMigrationService()
+        {
+            @Override
+            public ActionURL getSpecimenRequestEventDownloadURL(SpecimenRequestEvent event, String name)
+            {
+                return SpecimenController2.getDownloadURL(event, name);
+            }
+
+            @Override
+            public ActionURL getOverviewURL(Container c)
+            {
+                return new ActionURL(OverviewAction.class, c);
+            }
+        });
+     }
 
     @Override
     protected void startupAfterSpringConfig(ModuleContext moduleContext)
